services/cdc-consumer/requirements.txt
kafka-python-ng==2.2.3
redis==5.0.8
prometheus-client==0.20.0
